add test 'Language & direction tests for Navbar' in App.test.tsx and add test languageSwitcher.test.tsx and i18n.test.tsx and moved to existing test from SignUpPage.spec.tsx
